+++ Thunderhead.embeddedframework/Thunderhead.xcframework/ios-arm64/Thunderhead.framework/Modules/module.modulemap
@@ -1,0 +1,19 @@
+framework module Thunderhead {
+  umbrella header "Thunderhead.h"
+  
+  export *
+  module * { export * }
+  
+  link framework "Accelerate"
+  link framework "CoreLocation"
+  link framework "Foundation"
+  link framework "LocalAuthentication"
+  link framework "CoreGraphics"
+  link framework "MessageUI"
+  link framework "SafariServices"
+  link framework "Security"
+  link framework "SystemConfiguration"
+  link framework "UIKit"
+  link framework "WebKit"
+  link "sqlite3"
+}
