--- Thunderhead.embeddedframework/Thunderhead.xcframework/ios-arm64_armv7/Thunderhead.framework/Headers/Thunderhead.h
@@ -1,19 +1,0 @@
-//
-//  Thunderhead.h
-//  Thunderhead
-//
-//  Copyright © 2020 Thunderhead. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for Thunderhead.
-FOUNDATION_EXPORT double ThunderheadVersionNumber;
-
-//! Project version string for Thunderhead.
-FOUNDATION_EXPORT const unsigned char ThunderheadVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <Thunderhead/PublicHeader.h>
-
-#import <Thunderhead/One.h>
-#import <Thunderhead/OneConfiguration.h>
