+++ Thunderhead.podspec
@@ -8,16 +8,16 @@
 
 Pod::Spec.new do |s|
   s.name                = "Thunderhead"
-  s.version             = "9.1.3"
+  s.version             = "10.0.0"
   s.summary             = "The Thunderhead SDK for iOS, for integrating the ONE Engagement Hub or Interaction Studio into your iOS application."
   s.authors             = { "Thunderhead ONE" => "onesupport@thunderhead.com"}
   s.homepage            = "https://github.com/thunderheadone/one-sdk-ios"
-  s.description         = "The Thunderhead SDK for iOS, for integrating ONE or Interaction Studio into your iOS application. The SDK supports iOS 9.0 and above."
+  s.description         = "The Thunderhead SDK for iOS, for integrating ONE or Interaction Studio into your iOS application. The SDK supports iOS 11.0 and above."
   s.license             = { :type => "Commercial", :text => "Copyright (c) 2020 Thunderhead One Inc. All rights reserved." }
   s.frameworks          = ["Accelerate", "CoreLocation", "Foundation", "LocalAuthentication", "CoreGraphics", "MessageUI", "SafariServices", "Security", "SystemConfiguration", "UIKit", "WebKit"]
   s.library             = "sqlite3"
   s.source = { :git => "https://github.com/thunderheadone/one-sdk-ios.git", :tag => "#{s.version}" }
-  s.platform            = :ios, "9.0"
+  s.platform            = :ios, "11.0"
   s.default_subspec     = 'Core'
   s.subspec 'Core' do |core|
     core.preserve_paths      = 'Thunderhead.embeddedframework/Thunderhead.xcframework'
